Factorize visualization test code and fix some tests
